very close to end.

+++ Characters/Archers/Archer.java
@@ -6,7 +6,7 @@ import Equipment.Armour;
 
 public class Archer extends Character {
 
-    public Archer(String name, double price, int attack, int defence, int health, int speed) {
+    public Archer(String name, int price, int attack, int defence, int health, int speed) {
         super(name, price, attack, defence, health, speed);
     }
