+++ Characters/Healers/Healer.java
@@ -3,7 +3,7 @@ package Characters.Healers;
 import Characters.Character;
 
 public class Healer extends Character {
-    public Healer(String name, double price, int attack, int defence, int health, int speed) {
+    public Healer(String name, int price, int attack, int defence, int health, int speed) {
         super(name, price, attack, defence, health, speed);
     }
 }
