+++ Characters/Knights/Knight.java
@@ -3,7 +3,7 @@ package Characters.Knights;
 import Characters.Character;
 
 public class Knight extends Character {
-    public Knight(String name, double price, int attack, int defence, int health, int speed) {
+    public Knight(String name, int price, int attack, int defence, int health, int speed) {
         super(name, price, attack, defence, health, speed);
     }
 }
