+++ Characters/Mages/Mage.java
@@ -3,7 +3,7 @@ package Characters.Mages;
 import Characters.Character;
 
 public class Mage extends Character {
-    public Mage(String name, double price, int attack, int defence, int health, int speed) {
+    public Mage(String name, int price, int attack, int defence, int health, int speed) {
         super(name, price, attack, defence, health, speed);
     }
 }
