+++ Characters/MythicalCreatures/MythicalCreature.java
@@ -5,7 +5,7 @@ import Characters.Character;
 public class MythicalCreature extends Character {
 
 
-    public MythicalCreature(String name, double price, int attack, int defence, int health, int speed) {
+    public MythicalCreature(String name, int price, int attack, int defence, int health, int speed) {
         super(name, price, attack, defence, health, speed);
     }
 }
